Modified unit tests and aligned with framework model
