+++ test/payments.js
@@ -2,7 +2,7 @@ process.env.NODE_ENV = 'test'
 
 const chai = require('chai')
 const chaiHttp = require('chai-http')
-const server = require('../../server')
+const server = require('../server')
 const should = chai.should()
 
 const paymentDetails = {
